fix(fido2): LargeBlob CBOR parse bug, extract Encapsulate ECDH helper
- Fix double SkipValue in ParseDecryptedBlob (corrupted parse on non-empty
  byte string keys — conditional skip based on whether key was consumed)
- Extract shared ECDH P-256 key agreement into PinUvAuthHelpers, eliminating
  ~40 lines of duplicated code between PinUvAuthProtocolV1/V2.Encapsulate
